+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.19.3
+- Re-added items to creative tabs